test(phases/functional/phase-frontmatter-process): reorganize functional tests
- skills/set-frontmatter/scripts/__tests__/functional/setfm-process.functional.spec.ts:
  Remove the old combined functional spec and type/category test cases.

- skills/set-frontmatter/scripts/phases/__tests__/functional/phase-frontmatter-process.functional.spec.ts:
  Add the phaseFrontmatter functional spec under the phases directory and
  update comments and imports to match the new module structure.
